app/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=0.1=cpu_0
blas=1.0=mkl
blosc=1.19.0=ha925a31_0
brotli=1.0.7=pypi_0
brotlipy=0.7.0=py37h4ab8f01_1000
bzip2=1.0.8=hfa6e2cd_2
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py37_0
cffi=1.14.0=py37ha419a9e_0
charls=2.1.0=h33f27b4_2
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.4.1=py_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py_2
cytoolz=0.10.1=py37hfa6e2cd_0
dash=1.13.3=pyh9f0ad1d_0
dash-bootstrap-components=0.10.2=pyh9f0ad1d_0
dash-canvas=0.1.0=pypi_0
dash-core-components=1.10.1=pyh9f0ad1d_0
dash-html-components=1.0.3=pyh9f0ad1d_0
dash-renderer=1.5.0=pyh9f0ad1d_0
dash-table=4.8.1=pyh9f0ad1d_0
dask-core=2.19.0=py_0
decorator=4.4.2=py_0
flask=1.1.2=pyh9f0ad1d_0
flask-compress=1.5.0=pyh9f0ad1d_0
freetype=2.10.2=hd328e21_0
future=0.18.2=py37hc8dfbb8_1
giflib=5.2.1=h2fa13f4_2
icc_rt=2019.0.0=h0cc432a_1
imagecodecs=2020.5.30=py37h92c78e3_1
imageio=2.8.0=py_0
intel-openmp=2019.4=245
itsdangerous=1.1.0=py_0
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.15.1=pypi_0
jpeg=9d=he774522_0
jxrlib=1.1=hfa6e2cd_2
kiwisolver=1.2.0=py37heaa310e_0
lcms2=2.11=he1115b7_0
libaec=1.0.4=he025d50_1
libmklml=2019.0.5=0
libpng=1.6.37=hfe6a214_1
libtiff=4.1.0=h885aae3_6
libwebp-base=1.1.0=hfa6e2cd_3
libzopfli=1.0.3=ha925a31_0
lz4-c=1.9.2=h62dcd97_1
markupsafe=1.1.1=py37h8055547_1
matplotlib-base=3.2.1=py37h911224e_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.1.0=py37h45dec08_0
mkl_random=1.1.0=py37h675688f_0
networkx=2.4=py_1
ninja=1.9.0=py37h74a9793_0
numpy=1.18.5=py37h6530119_0
numpy-base=1.18.5=py37hc3f5095_0
olefile=0.46=py_0
openjpeg=2.3.1=h57dd2e7_3
openssl=1.1.1g=he774522_0
pillow=7.1.2=py37hc826c6e_0
pip=20.1.1=py37_1
plotly=4.8.1=pyh9f0ad1d_0
pycparser=2.20=py_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.7.6=h60c2a47_2
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.5.1=py3.7_cuda101_cudnn7_0
pywavelets=1.1.1=py37h44b1f71_1
pyyaml=5.3.1=py37h8055547_0
retrying=1.3.3=py_2
scikit-image=0.17.2=py37h3bbf574_1
scikit-learn=0.23.1=pypi_0
scipy=1.5.0=py37h9439919_0
setuptools=47.3.1=py37_0
six=1.15.0=pyh9f0ad1d_0
snappy=1.1.8=ha925a31_2
sqlite=3.32.3=h2a8f88b_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.6.3=py_1
tk=8.6.10=hfa6e2cd_0
toolz=0.10.0=py_0
torchvision=0.6.1=py37_cu101
tornado=6.0.4=py37hfa6e2cd_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_2
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py37_0
wincertstore=0.2=py37_0
xz=5.2.5=h2fa13f4_0
yaml=0.2.5=he774522_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.4=h9f78265_3
